Menambahkan Tampilan Matakuliah di Menu Penilaian

+++ src/api/api.js
@@ -27,12 +27,16 @@ axiosInstance.interceptors.request.use(
 axiosInstance.interceptors.response.use(
   (response) => response,
   (error) => {
+    // Cek jika ada error response dengan status 401 (Unauthorized)
     if (error.response?.status === 401) {
-      localStorage.clear()
-      window.location.href = '/login'
+      // Pastikan error bukan dari halaman login itu sendiri untuk menghindari loop
+      if (!error.config.url.includes('/login')) {
+        localStorage.clear() // Hapus semua data dari local storage
+        window.location.href = '/login' // Arahkan pengguna ke halaman login
+      }
     }
+    // Kembalikan error agar bisa ditangani di tempat lain jika perlu
     return Promise.reject(error)
   }
 )
-
 export default axiosInstance
